+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update README and add logo
 - Make mass property editable
 - Fixed hierarchy path matching 
+- Add test case for state module
 
 ## [0.1.1] - 2026-01-28